requirements.txt
langchain
langchain-ollama
langchain-chroma
langchain-google-genai
langchain-community
chromadb
python-dotenv
pypdf
python-docx
docx2txt
flask
flask-cors
